fix(editorkit): let stacked editor layers show through each other
Converting maze-muncher to layers (games #1273) showed the stacked board
never actually stacks. Every layer sits in the same grid area, the active
one at opacity 1 and the rest at 0.28 — but `.editor-cell` paints an
opaque default background and a tile's `color` is the only override, so
the active layer covers its neighbours completely. The creator sees one
layer at a time, which is the one thing layers exist not to do.

A tile that declares no colour is now a hole: `.is-blank` clears the
background inside a layered board, so lower layers read through it. That
keeps the rule an author can learn in one line — paint what matters,
leave the rest colourless — and needs no contract change, since `color`
was already optional.

Muted layers move 0.28 -> 0.55. The old value was never visible through
an opaque layer above it, so it had never been judged against anything;
at 0.28 a dark underlay is effectively invisible once it can show.

Both halves are pinned: a CSS-contract assertion, and a component test
that gives one layer two coloured tiles and seven colourless ones, then
asserts seven holes and two painted cells. Confirmed the component test
fails without the `.is-blank` class before restoring it.

The layered tests move to their own file to stay inside the module-size
ratchet, matching how this panel's tests are already split by concern.
That drops EditorCollections.test.tsx from 595 lines to 495, and its
baseline is tightened to match rather than left loose.
