+++ eslint-rules/module-size-baseline.json
@@ -727,7 +727,7 @@
     "apps/web/src/styles.chipStrip.test.ts": 61,
     "apps/web/src/styles.codeSurfaceButtons.test.ts": 67,
     "apps/web/src/styles.css": 21679,
-    "apps/web/src/styles.editorSurface.test.ts": 37,
+    "apps/web/src/styles.editorSurface.test.ts": 43,
     "apps/web/src/styles.gamePageRemix.test.ts": 85,
     "apps/web/src/styles.gameTheater.test.ts": 91,
     "apps/web/src/styles.heroComposer.test.ts": 87,
@@ -842,9 +842,10 @@
     "apps/web/src/surfaces/studio/CreatorStudioView.tsx": 1192,
     "apps/web/src/surfaces/studio/diffLines.test.ts": 50,
     "apps/web/src/surfaces/studio/diffLines.ts": 46,
-    "apps/web/src/surfaces/studio/EditorCollections.test.tsx": 595,
+    "apps/web/src/surfaces/studio/EditorCollections.test.tsx": 495,
     "apps/web/src/surfaces/studio/editorJsonLiveDiff.test.ts": 127,
     "apps/web/src/surfaces/studio/editorJsonLiveDiff.ts": 88,
+    "apps/web/src/surfaces/studio/EditorLayers.test.tsx": 190,
     "apps/web/src/surfaces/studio/EditorPanel.surfaceMode.test.ts": 45,
     "apps/web/src/surfaces/studio/EditorPanel.test.tsx": 285,
     "apps/web/src/surfaces/studio/EditorPanel.tsx": 1090,
